Moved force purge test for local cache into linux only test file
